final-project/final-pjt-back/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.12
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==13.12.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.9
